requirements.txt
browser-use
langchain-groq
python-dotenv
asyncio
